Issue #21 MapApi splitted
